backend/requirements.txt
async-timeout==4.0.2
bottle==0.12.21
Deprecated==1.2.13
nanoid==2.0.0
packaging==21.3
pyparsing==3.0.9
python-dotenv==0.20.0
redis==4.3.3
wrapt==1.14.1
